+++ src/realm/group_shared.hpp
@@ -27,7 +27,10 @@
 #include <limits>
 #include <realm/util/features.h>
 #include <realm/util/thread.hpp>
-#include <realm/util/platform_specific_condvar.hpp>
+#ifndef _WIN32
+#include <realm/util/interprocess_condvar.hpp>
+#endif
+#include <realm/util/interprocess_mutex.hpp>
 #include <realm/group.hpp>
 #include <realm/handover_defs.hpp>
 #include <realm/impl/transact_log.hpp>
@@ -43,8 +46,8 @@ class WriteLogCollector;
 /// Thrown by SharedGroup::open() if the lock file is already open in another
 /// process which can't share mutexes with this process
 struct IncompatibleLockFile: std::runtime_error {
-    IncompatibleLockFile():
-        std::runtime_error("Incompatible lock file")
+    IncompatibleLockFile(const std::string& msg):
+        std::runtime_error("Incompatible lock file. " + msg)
     {
     }
 };
@@ -69,7 +72,7 @@ struct IncompatibleLockFile: std::runtime_error {
 /// stopped (end_read(), commit(), rollback()).
 ///
 /// If a transaction is active when the SharedGroup is destroyed, that
-/// transaction is implicitely terminated, either by a call to
+/// transaction is implicitly terminated, either by a call to
 /// end_read() or rollback().
 ///
 /// Two processes that want to share a database file must reside on
@@ -79,30 +82,30 @@ struct IncompatibleLockFile: std::runtime_error {
 /// Desired exception behavior (not yet fully implemented)
 /// ------------------------------------------------------
 ///
-///  - If any data access API function throws an unexpcted exception during a
+///  - If any data access API function throws an unexpected exception during a
 ///    read transaction, the shared group accessor is left in state "error
 ///    during read".
 ///
-///  - If any data access API function throws an unexpcted exception during a
+///  - If any data access API function throws an unexpected exception during a
 ///    write transaction, the shared group accessor is left in state "error
 ///    during write".
 ///
 ///  - If SharedGroup::begin_write() or SharedGroup::begin_read() throws an
-///    unexpcted exception, the shared group accessor is left in state "no
+///    unexpected exception, the shared group accessor is left in state "no
 ///    transaction in progress".
 ///
 ///  - SharedGroup::end_read() and SharedGroup::rollback() do not throw.
 ///
-///  - If SharedGroup::commit() throws an unexpcted exception, the shared group
+///  - If SharedGroup::commit() throws an unexpected exception, the shared group
 ///    accessor is left in state "error during write" and the transaction was
-///    not comitted.
+///    not committed.
 ///
 ///  - If SharedGroup::advance_read() or SharedGroup::promote_to_write() throws
-///    an unexpcted exception, the shared group accessor is left in state "error
+///    an unexpected exception, the shared group accessor is left in state "error
 ///    during read".
 ///
 ///  - If SharedGroup::commit_and_continue_as_read() or
-///    SharedGroup::rollback_and_continue_as_read() throws an unexpcted
+///    SharedGroup::rollback_and_continue_as_read() throws an unexpected
 ///    exception, the shared group accessor is left in state "error during
 ///    write".
 ///
@@ -136,14 +139,14 @@ public:
         durability_Async    ///< Not yet supported on windows.
     };
 
-    /// \brief Same as calling the corrsponding version of open() on a instance
+    /// \brief Same as calling the corresponding version of open() on a instance
     /// constructed in the unattached state.
     explicit SharedGroup(const std::string& file, bool no_create = false,
                          DurabilityLevel durability = durability_Full,
                          const char* encryption_key = nullptr,
                          bool allow_file_format_upgrade = true);
 
-    /// \brief Same as calling the corrsponding version of open() on a instance
+    /// \brief Same as calling the corresponding version of open() on a instance
     /// constructed in the unattached state.
     explicit SharedGroup(Replication& repl,
                          DurabilityLevel durability = durability_Full,
@@ -177,7 +180,7 @@ public:
     /// specify the same durability level, otherwise an exception will be
     /// thrown.
     ///
-    /// If \a allow_file_format_ugrade is set to `true`, this function will
+    /// If \a allow_file_format_upgrade is set to `true`, this function will
     /// automatically upgrade the file format used in the specified Realm file
     /// if necessary (and if it is possible). In order to prevent this, set \a
     /// allow_upgrade to `false`.
@@ -207,7 +210,7 @@ public:
               const char* encryption_key = nullptr, bool allow_file_format_upgrade = true);
 
     /// Open this group in replication mode. The specified Replication instance
-    /// must remain in exixtence for as long as the SharedGroup.
+    /// must remain in existence for as long as the SharedGroup.
     void open(Replication&, DurabilityLevel = durability_Full,
               const char* encryption_key = nullptr, bool allow_file_format_upgrade = true);
 
@@ -246,7 +249,8 @@ public:
     /// NOTE:
     /// "changed" means that one or more commits has been made to the database
     /// since the SharedGroup (on which wait_for_change() is called) last
-    /// started, committed, promoted or advanced a transaction.
+    /// started, committed, promoted or advanced a transaction. If the SharedGroup
+    /// has not yet begun a transaction, "changed" is undefined.
     ///
     /// No distinction is made between changes done by another process
     /// and changes done by another thread in the same process as the caller.
@@ -254,7 +258,6 @@ public:
     /// Has db been changed ?
     bool has_changed();
 
-#if !REALM_PLATFORM_APPLE
     /// The calling thread goes to sleep until the database is changed, or
     /// until wait_for_change_release() is called. After a call to wait_for_change_release()
     /// further calls to wait_for_change() will return immediately. To restore
@@ -267,7 +270,6 @@ public:
 
     /// re-enable waiting for change
     void enable_wait_for_change();
-#endif // !REALM_PLATFORM_APPLE
     // Transactions:
 
     using version_type = _impl::History::version_type;
@@ -351,7 +353,7 @@ public:
     /// transaction, which means that they can no longer be used to access the
     /// underlying objects.
     ///
-    /// Subordinate accessors that were detatched at the end of the previous
+    /// Subordinate accessors that were detached at the end of the previous
     /// read or write transaction will not be automatically reattached when a
     /// new transaction is initiated. The application must reobtain new
     /// accessors during a new transaction to regain access to the underlying
@@ -399,14 +401,14 @@ public:
     /// - The method will throw if called in unattached state.
     /// - The method will return false if other SharedGroups are accessing the database
     ///   in which case compaction is not done. This is not necessarily an error.
-    /// It will return true following succesful compaction.
+    /// It will return true following successful compaction.
     /// While compaction is in progress, attempts by other
     /// threads or processes to open the database will wait.
     /// Be warned that resource requirements for compaction is proportional to the amount
     /// of live data in the database.
-    /// Compaction works by writing the database contents to a temporary databasefile and
+    /// Compaction works by writing the database contents to a temporary database file and
     /// then replacing the database with the temporary one. The name of the temporary
-    /// file is formed by appending ".tmp_compaction_space" to the name of the databse
+    /// file is formed by appending ".tmp_compaction_space" to the name of the database
     ///
     /// FIXME: This function is not yet implemented in an exception-safe manner,
     /// therefore, if it throws, the application should not attempt to
@@ -446,7 +448,7 @@ public:
     ///   ConstSourcePayload::Stay.
     ///
     /// For all other (non-TableView) accessors, handover is done with payload copy,
-    /// since the payload is trival.
+    /// since the payload is trivial.
     ///
     /// Handover *without* payload is useful when you want to ship a tableview with its query for
     /// execution in a background thread. Handover with *payload move* is useful when you want to
@@ -473,7 +475,7 @@ public:
     /// a tableview dependent upon another tableview and using MutableSourcePayload::Move,
     /// you are on thin ice!
     ///
-    /// On the importing side, the toplevel accessor being created during import takes ownership
+    /// On the importing side, the top-level accessor being created during import takes ownership
     /// of all other accessors (if any) being created as part of the import.
 
     /// Type used to support handover of accessors between shared groups.
@@ -530,20 +532,24 @@ private:
     bool m_wait_for_change_enabled;
     std::string m_lockfile_path;
     std::string m_db_path;
+    std::string m_coordination_dir;
     const char* m_key;
     TransactStage m_transact_stage;
     util::Mutex m_handover_lock;
+    util::InterprocessMutex m_writemutex;
+    util::InterprocessMutex m_balancemutex;
+    util::InterprocessMutex m_controlmutex;
 #ifndef _WIN32
-    util::PlatformSpecificCondVar m_room_to_write;
-    util::PlatformSpecificCondVar m_work_to_do;
-    util::PlatformSpecificCondVar m_daemon_becomes_ready;
-    util::PlatformSpecificCondVar m_new_commit_available;
+    util::InterprocessCondVar m_room_to_write;
+    util::InterprocessCondVar m_work_to_do;
+    util::InterprocessCondVar m_daemon_becomes_ready;
+    util::InterprocessCondVar m_new_commit_available;
 #endif
 
     void do_open(const std::string& file, bool no_create, DurabilityLevel, bool is_backend,
                  const char* encryption_key, bool allow_file_format_upgrade);
 
-    // Ring buffer managment
+    // Ring buffer management
     bool        ringbuf_is_empty() const noexcept;
     size_t ringbuf_size() const noexcept;
     size_t ringbuf_capacity() const noexcept;
@@ -556,7 +562,7 @@ private:
     void        ringbuf_put(const ReadCount& v);
     void        ringbuf_expand();
 
-    /// Grab a read lock on the spapshot associated with the specified
+    /// Grab a read lock on the snapshot associated with the specified
     /// version. If `version_id == VersionID()`, a read lock will be grabbed on
     /// the latest available snapshot. Fails if the snapshot is no longer
     /// available.
@@ -864,7 +870,7 @@ std::unique_ptr<SharedGroup::Handover<T>> SharedGroup::export_for_handover(const
     std::unique_ptr<Handover<T>> result(new Handover<T>());
     // Implementation note:
     // often, the return value from clone will be T*, BUT it may be ptr to some base of T
-    // instead, so we must cast it to T*. This is alway safe, because no matter the type,
+    // instead, so we must cast it to T*. This is always safe, because no matter the type,
     // clone() will clone the actual accessor instance, and hence return an instance of the
     // same type.
     result->clone.reset(dynamic_cast<T*>(accessor.clone_for_handover(result->patch, mode).release()));
@@ -1041,7 +1047,7 @@ inline bool SharedGroup::do_advance_read(O* observer, VersionID version_id, _imp
         observer->parse_complete(); // Throws
     }
 
-    // The old read lock must be retaind for as long as the change history is
+    // The old read lock must be retained for as long as the change history is
     // accessed (until Group::advance_transact() returns). This ensures that the
     // oldest needed changeset remains in the history, even when the history is
     // implemented as a separate unversioned entity outside the Realm (i.e., the
@@ -1128,6 +1134,11 @@ public:
     static int get_file_format_version(const SharedGroup& sg) noexcept
     {
         return sg.get_file_format_version();
+    }
+
+    static SharedGroup::version_type get_version_of_latest_snapshot(SharedGroup& sg)
+    {
+        return sg.get_version_of_latest_snapshot();
     }
 
     static SharedGroup::version_type get_version_of_bound_snapshot(const SharedGroup& sg) noexcept