+++ src/realm/query_expression.hpp
@@ -718,8 +718,8 @@ struct NullableVector
         m_first[index] = m_null;
     }
 
-    template <typename Type = t_storage> typename std::enable_if<
-        std::is_same<Type, int64_t>::value, void>::type
+    template <typename Type = t_storage>
+    typename std::enable_if<std::is_same<Type, int64_t>::value, void>::type
     set(size_t index, t_storage value)
     {
         REALM_ASSERT((std::is_same<t_storage, int64_t>::value));
@@ -735,8 +735,9 @@ struct NullableVector
         m_first[index] = value;
     }
 
-    template <typename Type = T> typename std::enable_if
-        <realm::is_any<Type, float, double, DateTime, BinaryData, StringData, NewDate, null>::value, void>::type
+    template <typename Type = T>
+    typename std::enable_if<realm::is_any<Type, float, double, DateTime, BinaryData, StringData, null>::value,
+                            void>::type
     set(size_t index, t_storage value) {
         m_first[index] = value;
     }
@@ -1655,7 +1656,7 @@ template <>
 class Columns<StringData> : public SimpleColumn<StringData> {
 public:
     using SimpleColumn::SimpleColumn;
-    std::unique_ptr<Subexpr> clone(QueryNodeHandoverPatches* np = nullptr) const override
+    std::unique_ptr<Subexpr> clone(QueryNodeHandoverPatches* = nullptr) const override
     {
         return make_subexpr<Columns<StringData>>(*this);
     }