+++ test/test-realm/test-realm.cpp
@@ -10,16 +10,6 @@ using namespace realm;
 
 namespace {
 
-// Get and Set are too fast (50ms/M) for normal 64-bit rand*rand*rand*rand*rand (5-10ms/M)
-uint64_t rand2()
-{
-    static int64_t seed = 2862933555777941757ULL;
-    static int64_t seed2 = 0;
-    seed = (2862933555777941757ULL * seed + 3037000493ULL);
-    seed2++;
-    return seed * seed2 + seed2;
-}
-
 REALM_TABLE_1(IntegerTable,
                 first, Int)